+++ examples/gltf_scenes.py
@@ -84,7 +84,14 @@ class CubeModel(CameraWindow):
             projection_matrix=self.camera.projection.matrix,
             camera_matrix=camera_matrix,
             children=True,
+            color=(0.75, 0.75, 0.75),
         )
+
+        # self.scene.draw_wireframe(
+        #     projection_matrix=self.camera.projection.matrix,
+        #     camera_matrix=camera_matrix,
+        #     color=(1, 1, 1, 1),
+        # )
 
 
 if __name__ == '__main__':