+++ moderngl_window/integrations/imgui.py
@@ -2,7 +2,7 @@ import ctypes
 
 import imgui
 import moderngl
-from imgui.integrations.opengl import BaseOpenGLRenderer
+from imgui.integrations.base import BaseOpenGLRenderer
 
 
 class ModernglWindowMixin:
@@ -40,10 +40,10 @@ class ModernglWindowMixin:
             self.io.mouse_down[0] = 1
 
         if self.wnd.mouse_states.middle:
-            self.io.mouse_down[1] = 1
+            self.io.mouse_down[2] = 1
 
         if self.wnd.mouse_states.right:
-            self.io.mouse_down[2] = 1
+            self.io.mouse_down[1] = 1
 
     def mouse_scroll_event(self, x_offset, y_offset):
         self.io.mouse_wheel = y_offset
@@ -55,10 +55,10 @@ class ModernglWindowMixin:
             self.io.mouse_down[0] = 1
 
         if button == self.wnd.mouse.middle:
-            self.io.mouse_down[1] = 1
+            self.io.mouse_down[2] = 1
 
         if button == self.wnd.mouse.right:
-            self.io.mouse_down[2] = 1
+            self.io.mouse_down[1] = 1
 
     def mouse_release_event(self, x: int, y: int, button: int):
         self.io.mouse_pos = self._mouse_pos_viewport(x, y)
@@ -67,10 +67,10 @@ class ModernglWindowMixin:
             self.io.mouse_down[0] = 0
 
         if button == self.wnd.mouse.middle:
-            self.io.mouse_down[1] = 0
+            self.io.mouse_down[2] = 0
 
         if button == self.wnd.mouse.right:
-            self.io.mouse_down[2] = 0
+            self.io.mouse_down[1] = 0
 
     def unicode_char_entered(self, char):
         io = imgui.get_io()
@@ -111,24 +111,37 @@ class ModernGLRenderer(BaseOpenGLRenderer):
         self._vertex_buffer = None
         self._index_buffer = None
         self._vao = None
+        self._textures = {}
         self.wnd = kwargs.get('wnd')
-        self.ctx = self.wnd.ctx
-
-        if not self.wnd:
-            raise ValueError('Missing window reference')
+        self.ctx = self.wnd.ctx if self.wnd and self.wnd.ctx else kwargs.get('ctx')
 
         if not self.ctx:
-            raise ValueError('Missing moderngl contex')
+            raise ValueError('Missing moderngl context')
+
+        assert isinstance(self.ctx, moderngl.context.Context)
 
         super().__init__()
+
+        if 'display_size' in kwargs:
+            self.io.display_size = kwargs.get('display_size')
+
+    def register_texture(self, texture: moderngl.Texture):
+        """Make the imgui renderer aware of the texture"""
+        self._textures[texture.glo] = texture
+
+    def remove_texture(self, texture: moderngl.Texture):
+        """Remove the texture from the imgui renderer"""
+        del self._textures[texture.glo]
 
     def refresh_font_texture(self):
         width, height, pixels = self.io.fonts.get_tex_data_as_rgba32()
 
         if self._font_texture:
+            self.remove_texture(self._font_texture)
             self._font_texture.release()
 
         self._font_texture = self.ctx.texture((width, height), 4, data=pixels)
+        self.register_texture(self._font_texture)
         self.io.fonts.texture_id = self._font_texture.glo
         self.io.fonts.clear_tex_data()
 
@@ -185,6 +198,16 @@ class ModernGLRenderer(BaseOpenGLRenderer):
 
             idx_pos = 0
             for command in commands.commands:
+                texture = self._textures.get(command.texture_id)
+                if texture is None:
+                    raise ValueError((
+                        "Texture {} is not registered. Please add to renderer using "
+                        "register_texture(..). "
+                        "Current textures: {}".format(command.texture_id, list(self._textures))
+                    ))
+
+                texture.use(0)
+
                 x, y, z, w = command.clip_rect
                 self.ctx.scissor = int(x), int(fb_height - w), int(z - x), int(w - y)
                 self._vao.render(moderngl.TRIANGLES, vertices=command.elem_count, first=idx_pos)