+++ moderngl_window/scene/material.py
@@ -37,13 +37,13 @@ class MaterialTexture:
 
 class Material:
     """Generic material"""
-    def __init__(self, name):
+    def __init__(self, name: str = None):
         """Initialize material.
 
         Args:
             name (str): Name of the material
         """
-        self._name = name
+        self._name = name or "default"
         self._color = (1.0, 1.0, 1.0, 1.0)
         self._mat_texture = None
         self._double_sided = True
@@ -83,6 +83,11 @@ class Material:
     @double_sided.setter
     def double_sided(self, value):
         self._double_sided = value
+
+    def release(self):
+        if self._mat_texture:
+            if self._mat_texture.texture:
+                self._mat_texture.texture.release()
 
     def __str__(self) -> str:
         return "<Material {}>".format(self.name)