+++ moderngl_window/scene/mesh.py
@@ -64,8 +64,18 @@ class Mesh:
         program["m_cam"].write(cam_matrix)
         program["bb_min"].write(self.bbox_min.astype('f4').tobytes())
         program["bb_max"].write(self.bbox_max.astype('f4').tobytes())
-        program["color"].value = (0.75, 0.75, 0.75)
         vao.render(program)
+
+    def draw_wireframe(self, proj_matrix, model_matrix, program):
+        """Render the mesh as wireframe.
+
+            proj_matrix: Projection matrix
+            model_matrix: View/model matrix
+            program: The moderngl.Program rendering the wireframe
+        """
+        program["m_proj"].write(proj_matrix)
+        program["m_model"].write(model_matrix)
+        self.vao.render(program)
 
     def add_attribute(self, attr_type, name, components):
         """